third_party/blink/web_tests/external/wpt/html/infrastructure/safe-passing-of-structured-data/structured-cloning-error-extra-expected.txt

This is a testharness.js-based test.
FAIL Throwing name getter fails serialization assert_throws_exactly: function "() => {
    worker.postMessage(badError);
  }" did not throw
FAIL Errors sent across realms should preserve their type assert_equals: expected "TypeError" but got "Error"
Harness: the test ran to completion.
